+++ Adapter/ShopwareAdapter/DataProvider/Translation/TranslationDataProvider.php
@@ -9,9 +9,9 @@ use Shopware\Models\Article\Configurator\Option;
 use Shopware\Models\Property\Option as PropertyGroupModel;
 use Shopware\Models\Property\Value as PropertyValueModel;
 use Shopware\Models\Shop\Shop as ShopModel;
+use SystemConnector\IdentityService\Struct\Identity;
 use SystemConnector\TransferObject\Product\Property\Property;
 use SystemConnector\TransferObject\Product\Property\Value\Value;
-use SystemConnector\ValueObject\Identity\Identity;
 
 class TranslationDataProvider implements TranslationDataProviderInterface
 {