+++ Components/Sepa/Plentymarkets/HandleSepaPaymentCommandHandler.php
@@ -7,13 +7,13 @@ use PlentymarketsAdapter\Client\ClientInterface;
 use PlentymarketsAdapter\PlentymarketsAdapter;
 use SystemConnector\IdentityService\Exception\NotFoundException;
 use SystemConnector\IdentityService\IdentityServiceInterface;
+use SystemConnector\IdentityService\Struct\Identity;
 use SystemConnector\ServiceBus\Command\CommandInterface;
 use SystemConnector\ServiceBus\Command\TransferObjectCommand;
 use SystemConnector\ServiceBus\CommandHandler\CommandHandlerInterface;
 use SystemConnector\ServiceBus\CommandType;
 use SystemConnector\TransferObject\Order\Order;
 use SystemConnector\TransferObject\Payment\Payment;
-use SystemConnector\ValueObject\Identity\Identity;
 
 class HandleSepaPaymentCommandHandler implements CommandHandlerInterface
 {